+++ analysis/tests.py
@@ -2269,7 +2269,6 @@ class TestChecks(TestCase):
         self.assertFalse(right_breakpoint_incorrect)
 
 
-
 class TestGnomad(TestCase):
     """
     Correctly handle displaying the Gnomad scores and links
@@ -2333,12 +2332,12 @@ class TestGnomad(TestCase):
     def test_gnomad_link_37(self):
         ''' link if build 37 '''
         self.variant_obj.genome_build=37
-        self.assertEqual(self.variant_instance_obj.gnomad_link(), 'https://gnomad.broadinstitute.org/variant/1:2345C>G?dataset=gnomad_r2_1')
+        self.assertEqual(self.variant_instance_obj.gnomad_link(), 'https://gnomad.broadinstitute.org/variant/1-2345-C-G?dataset=gnomad_r2_1')
 
     def test_gnomad_link_38(self):
         ''' link if build 38 '''
         self.variant_obj.genome_build=38
-        self.assertEqual(self.variant_instance_obj.gnomad_link(), 'https://gnomad.broadinstitute.org/variant/1:2345C>G?dataset=gnomad_r3')
+        self.assertEqual(self.variant_instance_obj.gnomad_link(), 'https://gnomad.broadinstitute.org/variant/1-2345-C-G?dataset=gnomad_r3')
 
     def test_gnomad_link_invalid_build(self):
         ''' value error should be thrown if not build 37 or 38 '''
@@ -2371,3 +2370,71 @@ class TestLIMSInitials(TestCase):
         ''' check will return false as initials already used by user '''
         initials_check, warning_message = lims_initials_check('ABC')
         self.assertEqual(initials_check, False)
+
+
+class TestPolyArtefactValidation(TestCase):
+    """
+    Checks that new Polys or Artefacts ran through Variant validator API
+    return the correct warning message
+    """
+    def test_incorrect_input(self):
+        # Test for sense check of input values
+        test_cases = [
+            (('TRAINS', 657383, 'A', 'C', 'GRCh37'), 'TRAINS is not a chromosome - please correct. Do not include "chr" in this box.'),
+            (('3', 657383, 'TRAINS', 'T', 'GRCh37'), 'Ref must consist only of A, T, C, and G - please correct ref: TRAINS'),
+            (('3', 657383, 'A', 'TRAINS', 'GRCh37'), 'Alt must consist only of A, T, C, and G - please correct alt: TRAINS'),
+        ]
+        
+        for input_args, expected_warning in test_cases:
+            with self.subTest(input_args=input_args):
+                result = validate_variant(*input_args)
+                if result.startswith("HTTP Request failed"):
+                    self.skipTest("Error contacting external API")
+                else:
+                    self.assertEqual(result, expected_warning)
+    
+    def test_no_warning(self):
+        # Test for a valid poly/artefact
+        result = validate_variant('7', 140453136, 'A', 'T', 'GRCh37')
+        expected_warning = None
+        if result is not None and result.startswith("HTTP Request failed"):
+            self.skipTest("Error contacting external API")
+        else:
+            self.assertEqual(result, expected_warning)
+
+    def test_wrong_reference(self):
+        # Test for reference base provided not matching the reference genome
+        result = validate_variant('7', 140453136, 'T', 'A', 'GRCh37')
+        expected_warning = 'Variant Validator Warnings: NC_000007.13:g.140453136T>A: Variant reference (T) does not agree with reference sequence (A);'
+        if result.startswith("HTTP Request failed"):
+            self.skipTest("Error contacting external API")
+        else:
+            self.assertEqual(result, expected_warning)
+        
+    def test_outside_boundaries(self):
+        # Test for variants outside border of chromosome
+        result = validate_variant('1', 999999999, 'A', 'T', 'GRCh37')
+        expected_warning = 'Variant Validator Warnings: The specified coordinate is outside the boundaries of reference sequence NC_000001.10;'
+        if result.startswith("HTTP Request failed"):
+            self.skipTest("Error contacting external API")
+        else:
+            self.assertEqual(result, expected_warning)
+    
+    def test_no_transcripts_overlap(self):
+        # Test for intergenic variants
+        result = validate_variant('4', 12345678, 'G', 'C', 'GRCh37')
+        expected_warning = 'Variant Validator Warnings: None of the specified transcripts ([\"mane_select\"]) fully overlap the described variation in the genomic sequence. Try selecting one of the default options;'
+        if result.startswith("HTTP Request failed"):
+            self.skipTest("Error contacting external API")
+        else:
+            self.assertEqual(result, expected_warning)
+        
+    def test_unexpected_error(self):
+        # Test for unexpected json structure from unanticipated genomic features.
+        # This variant is in a pseudogene, hence the unexpected error.
+        result = validate_variant('1', 23456, 'G', 'A', 'GRCh37')
+        expected_warning = 'Unexpected Error, contact Bioinformatics'
+        if result.startswith("HTTP Request failed"):
+            self.skipTest("Error contacting external API")
+        else:
+            self.assertEqual(result, expected_warning)