+++ swgs/models.py
@@ -2,6 +2,7 @@ import random
 import string
 
 from django.db import models
+from django.contrib.auth.models import User
 from auditlog.registry import auditlog
 
 #####################
@@ -559,6 +560,18 @@ class AbstractVariantInstance(models.Model):
     """
     Abstract class for variant instance. Stores the fields common to germline and somatic instances
     """
+    
+    OUTCOME_CHOICES = (
+        ('G', 'Genuine'),
+        ('A', 'Artefact'),
+        ('N', 'Not Analysed')
+    )
+
+    STATUS_CHOICES = (
+        ('P','Pending'),
+        ('C','Complete'),
+    )
+    
     id = models.AutoField(primary_key=True)
     variant = models.ForeignKey("Variant", on_delete=models.CASCADE)
     patient_analysis = models.ForeignKey("PatientAnalysis", on_delete=models.CASCADE)
@@ -569,6 +582,8 @@ class AbstractVariantInstance(models.Model):
     max_splice_ai = models.DecimalField(max_digits=4, decimal_places=3, null=True, blank=True)
     gnomad_popmax_af = models.DecimalField(max_digits=4, decimal_places=3, null=True, blank=True)
     gnomad_nhomalt = models.IntegerField(null=True, blank=True)
+    status = models.CharField(max_length=1, choices=STATUS_CHOICES, default='P')
+    decision = models.CharField(max_length=1, choices=OUTCOME_CHOICES, null=True, blank=True)
     
     class Meta:
         abstract = True
@@ -661,6 +676,44 @@ class GermlineVariantInstance(AbstractVariantInstance):
             # otherwise the nearest variants are > 2bp away
             return False
         
+    def get_all_checks(self):
+        """
+        Get all checks carried out on the variant for displaying
+        """
+
+        checks = []
+
+        all_checks = GermlineIGVCheck.objects.filter(variant_instance__id = self.id).order_by('check_date')
+
+        for check in all_checks:
+
+            check_dict = {
+                'decision': check.get_decision_display(),
+                'user': check.user.username
+            } 
+
+            checks.append(check_dict)
+
+        return checks
+    
+    def update_status(self):
+        """
+        Update status if last two checks are matching
+        """
+
+        all_checks = GermlineIGVCheck.objects.filter(variant_instance__id = self.id).order_by('-check_date')
+
+        #if only one check, don't do anything, otherwise if the last two match, update status
+        if len(all_checks) > 1:
+            
+            if all_checks[0].decision == all_checks[1].decision and all_checks[0].user != all_checks[1].user:
+
+                #Don't want to complete if the decision is not analysed
+                if all_checks[0].decision != "N":
+                
+                    self.decision = all_checks[0].decision
+                    self.status = "C"
+                    self.save()        
 
 class SomaticVariantInstance(AbstractVariantInstance):
     """
@@ -725,3 +778,77 @@ class SomaticVariantInstance(AbstractVariantInstance):
             
             # otherwise the nearest variants are > 2bp away
             return False
+        
+    def get_all_checks(self):
+        """
+        Get all checks carried out on the variant for displaying
+        """
+
+        checks = []
+
+        all_checks = SomaticIGVCheck.objects.filter(variant_instance__id = self.id).order_by('check_date')
+
+        for check in all_checks:
+
+            check_dict = {
+                'decision': check.get_decision_display(),
+                'user': check.user.username
+            } 
+
+            checks.append(check_dict)
+
+        return checks
+
+    def update_status(self):
+        """
+        Update status if last two checks are matching
+        """
+
+        all_checks = SomaticIGVCheck.objects.filter(variant_instance__id = self.id).order_by('-check_date')
+
+        #if only one check, don't do anything, otherwise if the last two match and are by different users, update status
+        if len(all_checks) > 1:
+            
+            if all_checks[0].decision == all_checks[1].decision and all_checks[0].user != all_checks[1].user:
+
+                #Don't want to complete if the decision is not analysed
+                if all_checks[0].decision != "N":
+                
+                    self.decision = all_checks[0].decision
+                    self.status = "C"
+                    self.save() 
+
+class AbstractVariantChecks(models.Model):
+    """
+    Abstract class for variant checks. Stores the fields common to all checks
+    """
+
+    OUTCOME_CHOICES = (
+        ('G', 'Genuine'),
+        ('A', 'Artefact'),
+        ('N', 'Not Analysed'),
+    )
+
+    decision = models.CharField(max_length=1, choices = OUTCOME_CHOICES, blank=True, null=True)
+    user = models.ForeignKey('auth.User', on_delete=models.PROTECT, blank=True, null=True)
+    check_date = models.DateTimeField(blank=True, null=True)
+        
+    class Meta:
+        abstract = True
+
+class GermlineIGVCheck(AbstractVariantChecks):
+    """
+    IGV checks for germline variants
+    """
+    
+    variant_instance = models.ForeignKey(GermlineVariantInstance, on_delete=models.CASCADE)
+
+class SomaticIGVCheck(AbstractVariantChecks):
+    """
+    IGV checks for somatic variants
+    """
+
+    variant_instance = models.ForeignKey(SomaticVariantInstance, on_delete=models.CASCADE)
+
+
+